requirements.txt
tensorflow>=2.15.0
keras>=2.15.0
vit-keras>=0.1.2
streamlit>=1.28.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
